Made the syntax highlighting more flexible.
If a cheat sheet ends in an extension, we'll attempt to use that to choose
syntax highlighter.

If CHEATCOLORS is set, and no syntax hightilghter can be found from the
filename, we'll try to use the value of CHEATCOLORS to decide on a syntax
highlighter.

If CHEATCOLORS is set, but doesn't resolve to a syntax highlighter, we'll fall
back to 'markdown'.